requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
trafilatura>=1.8.0
feedparser>=6.0.0
googlenewsdecoder>=0.1.2
python-dotenv>=1.0.0
pyyaml>=6.0.1
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
markdown>=3.5
